Text<->Number conversion improved. Generic convertors implemented with tunable strings. Minor settings beautification.

+++ PluginInterface/INumberToText.cs
@@ -2,6 +2,6 @@
 {
     public interface INumberToText
     {
-        public abstract string ConvertNumberToString(int val);
+        public abstract string ConvertNumberToString(long val);
     }
 }
