+++ PluginInterface/NumberToTextEmpty.cs
@@ -2,7 +2,7 @@
 {
     public class NumberToTextEmpty : INumberToText
     {
-        public string ConvertNumberToString(int val)
+        public string ConvertNumberToString(long val)
         {
             return string.Empty;
         }
